Added Machine learning
Added Machine learning. It now learns from previously asked unknown questions. This way the more you ask and teach it, the smarter it becomes! *Todo* Add Command and answers to a seperate file to shrinkify the lines of code in the file and more easily add commands without the program forgetting them